Arb strategy added

+++ Sushiswap.py
@@ -47,7 +47,7 @@ class SushiswapUtils(object):
         amount_in_with_fee = amount_in*997
         numerator = amount_in_with_fee*reserve_out
         denominator = reserve_in*1000 + amount_in_with_fee
-        return int(numerator/denominator)
+        return float(numerator/denominator)
 
     @staticmethod
     def get_amount_in(amount_out, reserve_in, reserve_out):
@@ -63,7 +63,7 @@ class SushiswapUtils(object):
         assert reserve_in > 0 and reserve_out > 0
         numerator = reserve_in*amount_out*1000
         denominator = (reserve_out - amount_out)*997
-        return int(numerator/denominator + 1)
+        return float(numerator/denominator + 1)
 
 
 class SushiswapObject(object):
